+++ docroot/modules/custom/demo_umami_content/src/InstallHelper.php
@@ -8,6 +8,7 @@ use Drupal\Core\Extension\ModuleHandlerInterface;
 use Drupal\Core\Path\AliasManagerInterface;
 use Drupal\Core\State\StateInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\Yaml\Yaml;
 use Drupal\Component\Utility\Html;
 
 /**
@@ -85,7 +86,8 @@ class InstallHelper implements ContainerInjectionInterface {
       ->importArticles()
       ->importPressReleases()
       ->importPages()
-      ->importBlockContent();
+      ->importBlockContent()
+      ->importMedia();
   }
 
   /**
@@ -230,6 +232,55 @@ class InstallHelper implements ContainerInjectionInterface {
       $this->storeCreatedContentUuids($uuids);
       fclose($handle);
     }
+    return $this;
+  }
+
+  /**
+   * Creates media items.
+   *
+   * @return $this
+   */
+  protected function importMedia() {
+    $uuids = [];
+    $module_path = $this->moduleHandler->getModule('demo_umami_content')->getPath();
+
+    $media = Yaml::parse(file_get_contents($module_path . '/default_content/media.yml'));
+
+    foreach ($media as $media_item) {
+      $values = [
+        'bundle' => 'image',
+        'name' => $media_item['name'],
+        'moderation_state' => 'published',
+      ];
+      $path = $module_path . '/default_content/images/' . $media_item['filename'];
+      $values['field_media_image'] = [
+        'target_id' => $this->createFileEntity($path),
+        'alt' => $media_item['alt'],
+      ];
+
+      $media = $this->entityTypeManager->getStorage('media')->create($values);
+      $media->save();
+      $uuids[$media->uuid()] = 'media';
+
+      // Save a hero crop for each of the images so that they can be easily
+      // used as a hero image on Press Releases.
+      $crop_values = [
+        'type' => 'landscape_hero',
+        'uri' => 'public://' . $media_item['filename'],
+        'height' => $media_item['crop']['height'],
+        'width' => $media_item['crop']['width'],
+        'x' => $media_item['crop']['x'],
+        'y' => $media_item['crop']['y'],
+      ];
+      $crop = $this->entityTypeManager->getStorage('crop')->create($crop_values);
+      $crop->save();
+      $uuids[$media->uuid()] = 'crop';
+    }
+
+    $crop = $this->entityTypeManager->getStorage('crop')->create($crop_values);
+    $crop->save();
+
+    $this->storeCreatedContentUuids($uuids);
     return $this;
   }