+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added Image Media type
 * Fixed bug with Entity Browsers
 * Added Hero image to Press Releases and associated crop type.
+* Create default media image content on install